Add files via upload
Hi guys, I have almost completed mine. And I have made some css changes today to make the page responsive. As for the result gallery, it can now change it's display as the width of the screen changes. But I don't know why for the vote page the single image div and the button's font size don't have any reaction to my setting it's width and height when screen width is less than 783px. I will fix it later. And I am thinking of giving the gallery a beautiful frame(if time permitted).